src/mysticscape/core/pipeline.py
"""
Core rendering pipeline and plugin system for Mystic Scape
"""

import os
import sys
import json
import importlib
import logging
from enum import Enum
from typing import Dict, List, Optional, Any
from pathlib import Path
import numpy as np
import OpenGL.GL as gl
import pyusd
from abc import ABC, abstractmethod

class RenderEngine(Enum):
    HYPERION = "hyperion"
    RENDERMAN = "renderman"
    VRAY = "vray"
    UNI_PATH_TRACER = "uni_path_tracer"
    BI_PATH_TRACER = "bi_path_tracer"
    PIXAR_UNIFIED = "pixar_unified"

class PluginType(Enum):
    RENDERER = "renderer"
    MODELING = "modeling"
    ANIMATION = "animation"
    SIMULATION = "simulation"
    PARTICLES = "particles"
    VOLUMETRICS = "volumetrics"
    COMPOSITING = "compositing"
    CUSTOM = "custom"

class Plugin(ABC):
    def __init__(self, name: str, version: str, plugin_type: PluginType):
        self.name = name
        self.version = version
        self.type = plugin_type
        self.enabled = False
        self.settings = {}

    @abstractmethod
    def initialize(self) -> bool:
        pass

    @abstractmethod
    def shutdown(self) -> None:
        pass

class PluginManager:
    def __init__(self):
        self.plugins: Dict[str, Plugin] = {}
        self.plugin_paths: List[Path] = []
        self.logger = logging.getLogger(__name__)

    def add_plugin_path(self, path: Path):
        """Add a directory to search for plugins"""
        if path.exists() and path.is_dir():
            self.plugin_paths.append(path)
            sys.path.append(str(path))

    def load_plugin(self, plugin_path: Path) -> Optional[Plugin]:
        """Load a plugin from file"""
        try:
            spec = importlib.util.spec_from_file_location(
                plugin_path.stem, plugin_path
            )
            if spec and spec.loader:
                module = importlib.util.module_from_spec(spec)
                spec.loader.exec_module(module)
                if hasattr(module, 'create_plugin'):
                    return module.create_plugin()
        except Exception as e:
            self.logger.error(f"Failed to load plugin {plugin_path}: {str(e)}")
        return None

    def scan_plugins(self):
        """Scan for available plugins"""
        for path in self.plugin_paths:
            for plugin_file in path.glob("*.py"):
                if plugin := self.load_plugin(plugin_file):
                    self.plugins[plugin.name] = plugin

class Pipeline:
    def __init__(self):
        self.plugin_manager = PluginManager()
        self.current_engine = RenderEngine.HYPERION
        self.logger = logging.getLogger(__name__)
        self.initialize_pipeline()

    def initialize_pipeline(self):
        """Initialize the rendering pipeline"""
        # Add default plugin paths
        plugin_dir = Path(__file__).parent / "plugins"
        self.plugin_manager.add_plugin_path(plugin_dir)
        
        # Initialize OpenUSD
        self.init_usd()
        
        # Setup render engines
        self.setup_render_engines()

    def init_usd(self):
        """Initialize OpenUSD workflow"""
        try:
            self.stage = pyusd.Stage.CreateInMemory()
            self.logger.info("OpenUSD initialized successfully")
        except Exception as e:
            self.logger.error(f"Failed to initialize OpenUSD: {str(e)}")

    def setup_render_engines(self):
        """Setup available render engines"""
        self.render_engines = {
            RenderEngine.HYPERION: self.create_hyperion_renderer(),
            RenderEngine.RENDERMAN: self.create_renderman_renderer(),
            RenderEngine.VRAY: self.create_vray_renderer(),
            RenderEngine.UNI_PATH_TRACER: self.create_uni_path_tracer(),
            RenderEngine.BI_PATH_TRACER: self.create_bi_path_tracer(),
            RenderEngine.PIXAR_UNIFIED: self.create_pixar_unified_renderer()
        }

    def set_render_engine(self, engine: RenderEngine):
        """Switch active render engine"""
        if engine in self.render_engines:
            self.current_engine = engine
            self.logger.info(f"Switched to {engine.value} renderer")
            return True
        return False

    def create_hyperion_renderer(self):
        """Create Hyperion renderer instance"""
        # TODO: Implement Hyperion renderer
        pass

    def create_renderman_renderer(self):
        """Create RenderMan renderer instance"""
        # TODO: Implement RenderMan renderer
        pass

    def create_vray_renderer(self):
        """Create V-Ray renderer instance"""
        # TODO: Implement V-Ray renderer
        pass

    def create_uni_path_tracer(self):
        """Create Uni-Directional Path Tracer"""
        # TODO: Implement Uni-Directional Path Tracer
        pass

    def create_bi_path_tracer(self):
        """Create Bi-Directional Path Tracer"""
        # TODO: Implement Bi-Directional Path Tracer
        pass

    def create_pixar_unified_renderer(self):
        """Create Pixar Unified renderer instance"""
        # TODO: Implement Pixar Unified renderer
        pass

    def execute_query(self, query: str, parallel: bool = True):
        """Execute a query in the pipeline"""
        # TODO: Implement query execution system
        pass

    def create_particle_system(self, params: Dict[str, Any]):
        """Create a particle system"""
        # TODO: Implement particle system
        pass

    def create_volumetrics(self, params: Dict[str, Any]):
        """Create volumetric effects"""
        # TODO: Implement volumetrics
        pass

    def send_support_email(self, subject: str, message: str):
        """Send support email"""
        import smtplib
        from email.mime.text import MIMEText
        
        support_email = "wambuiraymnd03@gmail.com"
        try:
            msg = MIMEText(message)
            msg['Subject'] = subject
            msg['From'] = "support@mysticscape.com"
            msg['To'] = support_email
            
            # TODO: Configure SMTP settings
            # server = smtplib.SMTP('smtp.mysticscape.com')
            # server.send_message(msg)
            # server.quit()
            
            self.logger.info(f"Support email sent: {subject}")
        except Exception as e:
            self.logger.error(f"Failed to send support email: {str(e)}")

    def load_plugin(self, plugin_path: str):
        """Load a plugin into the pipeline"""
        path = Path(plugin_path)
        if plugin := self.plugin_manager.load_plugin(path):
            if plugin.initialize():
                self.logger.info(f"Successfully loaded plugin: {plugin.name}")
                return True
        return False
