requirements.txt
numpy>=1.21.0
OpenGL>=3.1.6
PyQt6>=6.4.0
PySide6>=6.4.0
scipy>=1.7.0
opencv-python>=4.6.0
torch>=1.10.0
pyopengl>=3.1.6
pillow>=9.0.0
pyrender>=0.1.45
trimesh>=3.9.0
pyglet>=2.0.0
matplotlib>=3.5.0
networkx>=2.6.0
